==> Performance_benchmarks/Bulk_simulations/Dmelanogaster_dataset/2_Dmelanogaster_DTU.Rmd <==
---
title: "2_Dmelanogaster_DTU"
author: "Jeroen Gilis"
date: "05/11/2020"
output: html_document
---

**In order to run this script (2_Dmelanogaster_DTU.Rmd), the dataset Dmelanogaster_benchmark_datasets_count.Rdata (or, alternatively, Dmelanogaster_benchmark_datasets_scaledTPM.Rdata) is required.** This file can either be generated with the 1_Dmelanogaster_prepare.Rmd script or downloaded from Zenodo.

Here we run the DTU analyses for 7 DTU methods on both Dmelanogaster benchmark datasets. This code runs only a couple of minutes on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime was attributed to the DRIMSeq analysis.

**If you do not want to run this script, its output can also be downloaded from Zenodo: Dmelanogaster_DTU_results_count.Rdata (or, alternatively, Dmelanogaster_DTU_results_scaledTPM.Rdata)**

```{r setup, include=FALSE, echo=FALSE}
require("knitr")
opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load DTU methods

Source file that allows running all seven DTU methods that were assessed in this paper: satuRn, DoubleExpSeq, limma diffsplice, edgeRdiffsplice, DEXSeq, DRIMseq and NBSplice.

```{r, message=FALSE, warning=FALSE}
### load libraries
library(edgeR)
library(limma)
library(DEXSeq)
library(DRIMSeq)
library(DoubleExpSeq)
library(NBSplice)
library(satuRn)
library(doMC)

source(file="./Performance_benchmarks/DTU_methods.R")
```

# Set up parallel execution

Note that the optimal parameter setting for parallelizing the DTU analyses depends on your specific computing system.

```{r}
if(TRUE) {
    nrCores <- 2

    if(nrCores != 1) {
        doParallel <- TRUE
        doProgress <- 'none'

        registerDoMC(cores = nrCores)
    } else {
        doParallel <- FALSE
        doProgress <- 'text'
    }
}
```

# Run the analysis for all methods on both Dmelanogaster datasets

Run the DTU analyses for 7 DTU methods on both Dmelanogaster benchmark datasets.
This code runs a couple of minutes on a MacBook Pro 2018, processor; 2,3 GHz Quad-Core Intel Core i5, 16GB RAM. Most of this runtime to the analysis with DRIMSeq.

```{r}
### Load benchmark data
load(file="./Data/Dmelanogaster_benchmark_datasets_count.Rdata")
#load(file="./Data/Dmelanogaster_benchmark_datasets_scaledTPM.Rdata")

DmBenchmarkLenient <- list(DmBenchmarkLenient)
DmBenchmarkStringent <- list(DmBenchmarkStringent)

names(DmBenchmarkLenient)   <- paste0(names(DmBenchmarkLenient)  , 'filterLenient')
names(DmBenchmarkStringent) <- paste0(names(DmBenchmarkStringent), 'filterStringent')

DmBenchmarkData <- c(DmBenchmarkLenient,DmBenchmarkStringent)

### Run DTU analysis on benchmark data

print("start satuRn")

tStart <- Sys.time()
suppressWarnings(DmDtuBenchmark_satuRn <- plyr::llply(
    .data = DmBenchmarkData,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- satuRn_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)
        
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
))
difftime(Sys.time(), tStart)

print("start DoubleExpSeq")

tStart <- Sys.time()
DmDtuBenchmark_DoubleExpSeq <- plyr::llply(
    .data = DmBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DoubleExpSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design,
            quiet=FALSE)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start limma diffsplice")

tStart <- Sys.time()
DmDtuBenchmark_limmaDiffsplice <- plyr::llply(
    .data = DmBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- limma_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME
        )]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start edgeR_diffsplice")

tStart <- Sys.time()
DmDtuBenchmark_edgeRdiffsplice <- plyr::llply(
    .data = DmBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- edgeR_diffsplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DEXSeq")

tStart <- Sys.time()
DmDtuBenchmark_DEXSeq <- plyr::llply(
    .data = DmBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DEXSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start DRIMSeq")

tStart <- Sys.time()
DmDtuBenchmark_DRIMSeq <- plyr::llply(
    .data = DmBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- DRIMSeq_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design)

        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

print("start NBSplice")

tStart <- Sys.time()
DmDtuBenchmark_NBSplice <- plyr::llply(
    .data = DmBenchmarkData,
    .parallel = doParallel,
    .progress = doProgress,
    .inform = TRUE,
    .fun = function(localData) {
        ### Perform DTU analysis
        localRes <- NBSplice_DTU(
            countData  = localData$data,
            tx2gene    = localData$metaInfo,
            sampleData = localData$design
        )
        
        ### Massage
        localRes$gene_modified <- localData$metaInfo$gene_modified[match(
            localRes$TXNAME, localData$metaInfo$TXNAME)]

        ### Return result
        return(list(dtuAnalysis = localRes))
    }
)
difftime(Sys.time(), tStart)

### add method name to list names for easy post-analysis
names(DmDtuBenchmark_satuRn) <- paste0('satuRn_', names(DmDtuBenchmark_satuRn))
names(DmDtuBenchmark_DoubleExpSeq) <- paste0('DoubleExpSeq_', names(DmDtuBenchmark_DoubleExpSeq))
names(DmDtuBenchmark_edgeRdiffsplice) <- paste0('edgeR_diffsplice_', names(DmDtuBenchmark_edgeRdiffsplice))
names(DmDtuBenchmark_limmaDiffsplice) <- paste0('limma_diffsplice_', names(DmDtuBenchmark_limmaDiffsplice))
names(DmDtuBenchmark_DEXSeq) <- paste0('DEXSeq_', names(DmDtuBenchmark_DEXSeq))
names(DmDtuBenchmark_DRIMSeq) <- paste0('DRIMSeq_', names(DmDtuBenchmark_DRIMSeq))
names(DmDtuBenchmark_NBSplice) <- paste0('NBSplice_', names(DmDtuBenchmark_NBSplice))

### Save result
save(DmDtuBenchmark_satuRn,
    DmDtuBenchmark_DoubleExpSeq,
    DmDtuBenchmark_limmaDiffsplice,
    DmDtuBenchmark_edgeRdiffsplice,
    DmDtuBenchmark_DEXSeq,
    DmDtuBenchmark_DRIMSeq,
    DmDtuBenchmark_NBSplice,
    file="./Data/Dmelanogaster_DTU_results_count.Rdata")

# save(DmDtuBenchmark_satuRn,
#     DmDtuBenchmark_DoubleExpSeq,
#     DmDtuBenchmark_limmaDiffsplice,
#     DmDtuBenchmark_edgeRdiffsplice,
#     DmDtuBenchmark_DEXSeq,
#     DmDtuBenchmark_DRIMSeq,
#     DmDtuBenchmark_NBSplice,
#     file="./Data/Dmelanogaster_DTU_results_scaledTPM.Rdata")
```
